Add workflow Info.plist file
Copied this Info.plist from the latest release and then included by fix described in #61
Including the Info.plist also the workflow to be recreated from the GitHub repo and allows for tracking and observing the changes to it.  As well as giving users the ability to submit useful PRs!

Here's a diff of the only change I made:

*** info_old.plist	2021-08-19 13:07:01.000000000 -0400
--- info.plist	2021-11-23 22:19:29.000000000 -0500
***************
*** 247,251 ****
  	$hex = `./colorpick -startColor ffffff -mode 6`;
  	$action = str_replace('&lt;color&gt;', $hex, $action);
! 	return `osascript -e 'tell application "Alfred 3" to search "hue $action"'`;
  }

--- 247,251 ----
  	$hex = `./colorpick -startColor ffffff -mode 6`;
  	$action = str_replace('&lt;color&gt;', $hex, $action);
! 	return `osascript -e 'tell application "Alfred" to search "hue $action"'`;
  }